# openapi.yaml
openapi: 3.0.3
info:
  title: elrek-system_GO
  description: Ez lenni elrek-system_GO backend cuccli
  version: 1.0.0
#servers:
#  - url: 'yas'
paths:
...
  /users:
    get:
      description: Endpoint for user(s) listing
      parameters:
        - name: is_active
          in: query
          description: Filter users by is_active
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: Returns a list of users
          content:
            application/json:
              schema:
                $ref: './models/user.yaml#/components/schemas/UserList'
        default:
          description: Unexpected error, error is returned in message
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
      tags:
        - users
    post:
      description: Endpoint for user creation
      requestBody:
        description: User object that needs to be added to the database
        required: true
        content:
          application/json:
            schema:
              $ref: './models/user.yaml#/components/schemas/UserCreate'
      responses:
        200:
          description: User was created successfully
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
        default:
          description: Unexpected error, error is returned in message
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
      tags:
        - users
...
  /users/{id}:
    get:
      description: Endpoint for user listing
      parameters:
        - name: id
          in: path
          description: ID of the user to list
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns a user
          content:
            application/json:
              schema:
                $ref: './models/user.yaml#/components/schemas/UserResponse'
        default:
          description: Unexpected error, error is returned in message
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
      tags:
        - users
    patch:
      description: Endpoint for user update
      parameters:
        - name: id
          in: path
          description: ID of the user to update
          required: true
          schema:
            type: string
      requestBody:
        description: Attributes of user object that needs to be updated in the database
        required: true
        content:
          application/json:
            schema:
              $ref: './models/user.yaml#/components/schemas/UserUpdate'
      responses:
        200:
          description: User was updated successfully
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
        default:
          description: Unexpected error, error is returned in message
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
      tags:
        - users
    delete:
      description: Endpoint for user deletion
      parameters:
        - name: id
          in: path
          description: ID of the user to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: User was deleted successfully
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
        default:
          description: Unexpected error, error is returned in message
          content:
            application/json:
              schema:
                $ref: './models/responses.yaml#/components/schemas/MessageOnlyResponse'
      tags:
        - users
...